Remove old class and rename new to take its place
