+++ engine/table/src/main/java/io/deephaven/engine/util/AbstractScriptSession.java
@@ -95,7 +95,7 @@ public abstract class AbstractScriptSession<S extends AbstractScriptSession.Snap
         this.classCacheDirectory = classCacheDirectory;
 
         queryScope = new ScriptSessionQueryScope();
-        final QueryCompiler compilerContext = InMemoryQueryCompiler.create(classCacheDirectory);
+        final QueryCompiler compilerContext = QueryCompilerImpl.create(classCacheDirectory);
 
         executionContext = ExecutionContext.newBuilder()
                 .markSystemic()
